// src/api/auth/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface AuthApi {
	authChallenge: (request: Api.AuthChallengeRequest) => Promise<AuthChallengeResponse>
	authVerify: (request: Api.AuthVerifyRequest) => Promise<AuthVerifyResponse>
}

export type AuthChallengeResponse = AuthChallenge200Response | AuthChallengeDefaultResponse

export interface AuthChallenge200Response {
	status: 200
	body: Api.AuthChallengeResponse
	headers?: never
}

export interface AuthChallengeDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type AuthVerifyResponse = AuthVerify200Response | AuthVerifyDefaultResponse

export interface AuthVerify200Response {
	status: 200
	body: Api.AuthVerifyResponse
	headers?: never
}

export interface AuthVerifyDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}